+++ website_support/models/website_support_ticket.py
@@ -348,6 +348,9 @@ class WebsiteSupportTicket(models.Model):
         support_ticket_menu = self.env['ir.model.data'].sudo().get_object('website_support', 'website_support_ticket_menu')
         support_ticket_action = self.env['ir.model.data'].sudo().get_object('website_support', 'website_support_ticket_action')
 
+        #Add them as a follower to the ticket so they are aware of any internal notes
+        new_id.message_subscribe_users(user_ids=new_id.category.cat_user_ids.ids)
+
         for my_user in new_id.category.cat_user_ids:
             values = notification_template.generate_email(new_id.id)
             values['body_html'] = values['body_html'].replace("_ticket_url_", "web#id=" + str(new_id.id) + "&view_type=form&model=website.support.ticket&menu_id=" + str(support_ticket_menu.id) + "&action=" + str(support_ticket_action.id) ).replace("_user_name_",  my_user.partner_id.name)